requirements.txt
google-genai
deepgram-sdk
python-dotenv
aiohttp
websockets
